refactor(packages): consolidate the duplicated D1 REST transport into @kampus/d1-rest leaf (#941)
The D1 REST transport (makeD1Rest) was hand-copied across three packages —
preview-seed / fts-backfill / moderator-grant — so the latent meta.changes
defect (#940) had to be fixed three times and the copies had drifted
incidentally (fts-backfill's toRestParams passed null through; the other two
rejected it). Move the canonical transport into one true-leaf package and point
all three consumers at it, à la the #859 @kampus/db-schema consolidation.

- New @kampus/d1-rest leaf: makeD1Rest + makeD1RestFromEnv / d1RestLayerFromEnv,
  toRestParams / assertRestParam, D1RestConfig / D1RestServices. Deps only
  @distilled.cloud/cloudflare + effect (true leaf — no @kampus/web, so
  fts-backfill's @kampus/web prod-dep stays acyclic).
- Canonical impl is the null-rejecting variant (assertRestParam, #569) the two
  copies with explicit tests used; the meta.changes mapping (#937/#940) is now
  single-sourced here.
- Deleted packages/{preview-seed,fts-backfill,moderator-grant}/src/d1-rest.ts and
  moderator-grant's duplicated d1-rest.unit.test.ts; repointed every import.
- Transport contract tested once in d1-rest/src/index.unit.test.ts: meta.changes
  carried + defaults to 0, toRestParams null rejection, batch single-POST.
  Consumer-specific tests (seed statement-building, backfill batch) stay,
  repointed to the import.
- README for the leaf (what/why/leaf-cycle constraint).
