+++ packages/fts-backfill/src/index.ts
@@ -1,5 +1,5 @@
+export {d1RestLayerFromEnv, makeD1Rest, makeD1RestFromEnv} from "@kampus/d1-rest";
 export type {BackfillDb, BackfillReport, SourceRow} from "./backfill.ts";
 export {backfill, buildBackfillStatements, makeBackfillDb} from "./backfill.ts";
-export {d1RestLayerFromEnv, makeD1Rest, makeD1RestFromEnv} from "./d1-rest.ts";
 export type {BackfillSchema} from "./schema.ts";
 export {backfillSchema} from "./schema.ts";
