+++ packages/moderator-grant/src/bin.ts
@@ -15,10 +15,10 @@
  */
 import {CredentialsFromEnv} from "@distilled.cloud/cloudflare/Credentials";
 import {NodeRuntime, NodeServices} from "@effect/platform-node";
+import {makeD1Rest} from "@kampus/d1-rest";
 import {Config, Console, Data, Effect, Layer, Option} from "effect";
 import {Command, Flag} from "effect/unstable/cli";
 import * as FetchHttpClient from "effect/unstable/http/FetchHttpClient";
-import {makeD1Rest} from "./d1-rest.ts";
 import {listModerators, makeGrantDb, type Selector, setRole} from "./grant.ts";
 
 class SelectorRequired extends Data.TaggedError("SelectorRequired")<{message: string}> {}
